refactor(diagnostic): extract !platform group onto PlatformCommandsMixin
Move the platform_grp command group + all its subcommands off DiagnosticCog
onto PlatformCommandsMixin in cogs/diagnostic/platform_group.py. discord.py's
CogMeta collects commands across the cog MRO, so the !platform surface and its
single-cog registration are unchanged. diagnostic_cog.py drops 799 -> 260 LOC,
clearing the 800-LOC cog ceiling that blocked the next !platform subcommand.

Executes idea docs/ideas/diagnostic-cog-platform-group-extraction-2026-06-16.md.
Pure refactor, no behaviour change.
